% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_extent.R
\name{calc_extent_ew}
\alias{calc_extent_ew}
\title{Calculate East-West Extent}
\usage{
calc_extent_ew(v)
}
\arguments{
\item{v}{A SpatVector object representing a polygon.}
}
\value{
A numeric value representing the average east-west extent in meters.
}
\description{
Calculates the average east-west extent of a polygon based on its bounding box.
}
\examples{
\dontrun{
library(terra)
polygon <- vect("path/to/polygon.shp")
ew_extent <- calc_extent_ew(polygon)
}
}
