% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_extent.R
\name{calc_extent_ns}
\alias{calc_extent_ns}
\title{Calculate North-South Extent}
\usage{
calc_extent_ns(v)
}
\arguments{
\item{v}{A SpatVector object representing a polygon.}
}
\value{
A numeric value representing the average north-south extent in meters.
}
\description{
Calculates the average north-south extent of a polygon based on its geographic bounding box.
}
\examples{
library(terra)
coords <- cbind(c(0, 0, 1, 1, 0), c(0, 1, 1, 0, 0))
polygon <- vect(coords, type = "polygon", crs = "EPSG:4326")
ns_extent <- calc_extent_ns(polygon)
}
