% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_extent.R
\name{calc_extent_ew}
\alias{calc_extent_ew}
\title{Calculate East-West Extent}
\usage{
calc_extent_ew(v)
}
\arguments{
\item{v}{A SpatVector object representing a polygon.}
}
\value{
A numeric value representing the average east-west extent in meters.
}
\description{
Calculates the average east-west extent of a polygon based on its geographic bounding box.
}
\examples{
library(terra)
coords <- cbind(c(0, 0, 1, 1, 0), c(0, 1, 1, 0, 0))
polygon <- vect(coords, type = "polygon", crs = "EPSG:4326")
ew_extent <- calc_extent_ew(polygon)
}
