% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distant_points.R
\name{get_distant_points}
\alias{get_distant_points}
\title{Find Most Distant Points on Convex Hull}
\usage{
get_distant_points(v)
}
\arguments{
\item{v}{A SpatVector object representing a polygon.}
}
\value{
A list containing:
\item{start_ll}{SpatVector of the southernmost point}
\item{end_ll}{SpatVector of the northernmost point}
\item{distance}{Maximum distance in meters}
\item{bearing}{Geographic bearing from start to end in degrees}
}
\description{
Identifies the pair of most distant points on the convex hull of a polygon,
ordered from south to north.
}
\examples{
\dontrun{
library(terra)
polygon <- vect("path/to/polygon.shp")
distant_pts <- get_distant_points(polygon)
}
}
